Bug 1283646 - Generate one share for each type of group, r=ekr
Rather than unconditionally generating shares for P-256 and FFDHE-2048,
only generate EC or FFDHE shares if one of the corresponding types of
suites is enabled.  As it happens, it's just as easy at that point to
generate a share on the first enabled group rather than fixed values.

Note: When HelloRetryRequest is implemented, we can easily change this
to only generate FFDHE shares if EC is disabled.