Bug 1368623 - Allow perl to be msys or strawberry, r=franziskus
On windows, mozilla-build comes with perl.  The way that gyp ends up
invoking perl leads to perl receiving a filename argument that contains
backslashes.  You can address this, as you have suggested, by installing
strawberry perl and putting that on the path ahead of /bin, but it's more
software and unnecessary.  This little python wrapper avoids that problem.  It
works with both strawberry and the copy that comes with msys.

Differential Revision: https://nss-review.dev.mozaws.net/D327