Bug 1413787 - Remove lazy initialization, r=ekr
The only subtle change here is that I needed to move the call to
ssl_FilterSupportedGroups() from ssl3_InitState() to ssl3_config_match_init().
It is the only thing in ssl3_InitState that depends on configuration.  The rest
is pretty straightforward.

The other change of note is that a newly imported socket will have
ss->ssl3.hs.ws set to wait_client_hello (server) or idle_handshake (client).
Previously, the value was initialized to idle_handshake only.