Bug 1418944 - Quote CC/CXX variables passed to nspr r=franziskus
The Makefile doesn't put quotes around the assignment of CC variables when
calling the nspr configure (if building with nss_build_all).

This breaks e.g. if someone tries to pass flags in the CC (hacky, but sometimes
makes things easier) or if there are spaces in the path to the compiler. In any
case, quoting makes things cleaner and shouldn't have any downsides.