Bug 1423018 - Retain write polling when 0-RTT is enabled, r=ekr
This is a nasty one.  I was having trouble with tstclnt when testing against
other implementations.  It would hang.

I made similar changes to that for kazuho and picotls, where the input file is
read for every connection.  So, I did that here too and it also worked nicely.
In order to get 0-RTT working though, I needed to teach ssl_Poll about 0-RTT.
Luckily, there was code already there for that and it just needed a tweak.

The only thing I ran into here was that boringssl (the server I was using to
test against here), was too fast.  By the time we had written out the
ClientHello, it had produced a response and we would complete the handshake
before leaving the handshake loop in ssl3_Do1stHandshake().  That meant that we
would never actually send any 0-RTT data, either before or after this patch.
Adding a sleep(1) to the handshake in boringssl did the trick and I could show
that we can send data before the handshake completes.  Nothing really
actionable here unless you can think of ways to make our handshake more
performant.  Mostly just information.

Separately, people have noticed that tstclnt writes 0-RTT after the first round
trip.  That's a symptom of not retaining the poll on write.