Bug 1422688 - Fix for blocking mode reads of 0-RTT, r=ekr
We talked about this in an earlier iteration of Bug 1386475, but we didn't like
that the code checked whether there was early data twice.  Turns out that it is
necessary for blocking sockets.  If we return PR_WOULD_BLOCK_ERROR on a
blocking socket, the calling code rightfully objects.  As it turns out,
selfserv is an example of this.  It was breaking with boringssl.

I don't know why this wasn't an issue with other clients, but I suspect that it
is because very few implementations are doing exactly the same thing with 0-RTT
and timing is critical here.