Bug 1432431 - Land Cryptol/SAW specs for Poly1305 r=franziskus
Summary:
I wrote Cryptol and SAW specs for Poly1305 that are able to verify our
implementations for a single input/key pair. The 128-bit multiplication isn't
verifiable as-is. It takes forever, almost literally. We would need to take
the HACL* approach and gently guide Z3 to link the mathematical spec to the
optimized spec.

Let's not do this now, let's just land the specs so we have them in the repo.

Reviewers: franziskus

Reviewed By: franziskus

Bug #: 1432431

Differential Revision: https://phabricator.services.mozilla.com/D430