Bug 1483129 - TLS 1.3 RFC version, r=ekr
This retains the ability to negotiate draft versions of DTLS 1.3, but uses the
final RFC version for TLS 1.3.

This also refactors the handling of the downgrade sentinel.  As we've discovered
- to our dismay - some MitM boxes forward handshake messages when they
shouldn't.  This could result in triggering the downgrade sentinel.  I've done
two things here:

- The server always sets the sentinel.  It reduces the assumed version if it
  only supports a draft version though on the basis that the client might
  expect the full version.

- The client has a new option SSL_ENABLE_HELLO_DOWNGRADE_CHECK which is disabled
  by default.  The client will reject a handshake that appears to be a downgrade
  only when this is explicitly enabled.  The client will allow an apparent
  downgrade to TLS 1.2 if it is running a draft version of TLS 1.3.

The allowance for a draft version is now only effective for DTLS 1.3.

Tests for version downgrade have been updated and enabled.  These were rotten in
a few ways, but nothing dramatic.