Bug 1471985, make SECKEY_Get{Public,Private}KeyType consistent on RSA-PSS, r=rrelyea,fkiefer
Summary:
In bug 1413596, we changed SECKEY_GetPrivateKeyType() to return
rsaPssKey, if the private key is restricted to RSA-PSS when importing.
Although the intention of this change was to extend the certutil
output to provide more information about key types, it introduced
inconsistency with the existing code, as SECKEY_GetPublicKeyType()
still returns rsaKey.

This patch partially revert the change and determine the actual
(restricted) key type in a different way, using CERT_GetCertKeyType()
and PK11_GetCertFromPrivateKey().

Reviewers: rrelyea, franziskus

Reviewed By: franziskus

Subscribers: franziskus

Bug #: 1471985

Differential Revision: https://phabricator.services.mozilla.com/D1911