Bug 1434943 - Support for MSVC in build.sh, r=jcj
Summary:
This adds basic support for MSVC to build.sh.  It uses the registry and vswhere
(which is part of the standard mozilla-build setup now) to work out paths and
set them properly.  It's probably a little fragile, but it's better than the
shoestring and tape we have in builds right now.

I took the liberty of sanitizing the command-line options a little here.  Mostly
that is sorting them, but I also deprecated the -m32 option in favour of
specifying target architecture with -t.  That turned out to be a lot cleaner.

Reviewers: jcj

Reviewed By: jcj

Bug #: 1434943

Differential Revision: https://phabricator.services.mozilla.com/D5125