Bug 1531236 - Accessor for certificate DER, r=jcj
Summary:
Forgot to put this up.  This will make the neqo wrapper considerably more
hygenic.  Having to explode the entire CERTCertificate struct (which is public
and never should have been) into the FFI is a complete disaster.  Better to
treat it as opaque and use an accessor function.

Reviewers: jcj

Tags: #secure-revision

Bug #: 1531236

Differential Revision: https://phabricator.services.mozilla.com/D24129