Bug 1563078 - Set authKeyBits for delegated credentials, r=jcj
The delegated credentials patch left the channel info unmodified, which meant
that it reported the key strength of the end entity certificate and not the
delegated credential.  For a using application, this is problematic because it
can't access information about delegated credentials.  In this case, the only
omission was the strength of the key.

Firefox checks key strength for the entire certificate chain according to its
policies, but it also wants to apply the same sort of policy to the delegated
credential.  In particular, it wants to ensure that an RSA credential (which
shouldn't be used, but whatever...) has a long enough modulus, because the NSS
policy is less strict than the Firefox one.

To address this use case, SSLChannelInfo.authKeyBits is set to the length of the
delegated credential key when delegated credentials are in use.  This is
consistent with the definition of the parameter, but implies a different
understanding of its meaning when delegated credentials are enabled.

Differential Revision: https://phabricator.services.mozilla.com/D36699