Bug 1528666 - Correct resumption validation checks, r=jcj
We allowed cross-suite resumption before, but it didn't work.  This enables that
for clients.

As a secondary minor tweak, clients will no longer validate the availability of
a cipher suite based on their configured version range when attempting
resumption.  Instead, they will check whether the suite works for the version
in the session that they are attempting to resume.  In theory, this doesn't
change anything because the previous session should not have selected an
incompatible combination of version and cipher suite, but it's worth being extra
precise.

Differential Revision: https://phabricator.services.mozilla.com/D20119