Bug 1508776 - Remove unneeded refcounting from SFTKSession r=mt,kjacobs
SFTKSession objects are only ever actually destroyed at PK11 session closure,
as the session is always the final holder -- and asserting refCount == 1 shows
that to be true. Because of that, NSC_CloseSession can just call
`sftk_DestroySession` directly and leave `sftk_FreeSession` as a no-op to be
removed in the future.

Differential Revision: https://phabricator.services.mozilla.com/D47010