Bug 1600144 - Treat ClientHello with message_seq of 1 as a second ClientHello, r=kjacobs
Summary:
The logic that deals with stateless HelloRetryRequest in DTLS
allows this one-off increment to the message_seq field in case the server was
operating statelessly.  However, when it does, it should insist on the
ClientHello carrying a cookie; concretely, it should set the flag that says that
a HelloRetryRequest was sent, even if it doesn't currently remember that it sent
one. That is the only way that this condition could be met.

Differential Revision: https://phabricator.services.mozilla.com/D55210