Bug 1603438 - Fix native tools build failure due to lack of zlib include dir if external r=jcj
Add ZLIB_INCLUDE_DIR variable

On Linux platform[1], the build system forces to use zlib from the
system instead of compiling the one located intree.

The following error is raised when the zlib header is installed
somewhere else than in the default system include path:

	ssl3con.c:39:18: fatal error: zlib.h: No such file or directory
	 #include "zlib.h"

The same trick setup for sqlite include directory is reproduced for
zlib. The build system disallows in any manner to give arguments to the
compiler explicity.

The variable ZLIB_INCLUDE_DIR point to the directory where the zlib
header is located.

[1]:
https://hg.mozilla.org/projects/nss/file/NSS_3_33_BRANCH/coreconf/Linux.mk#l180
[2]:
https://hg.mozilla.org/projects/nss/file/NSS_3_33_BRANCH/lib/softoken/manifest.mn#l17