Bug 1629661 MPConfig calls in SSL initializes policy before NSS is initialized. r=mt
NSS has several config functions that multiprocess servers must call before NSS is initialized to set up shared memory caches between the processes. These functions call ssl_init(), which initializes the ssl policy. The ssl policy initialization, however needs to happen after NSS itself is initialized. Doing so before hand causes (in the best case) policy to be ignored by these servers, and crashes (in the worst case).

Instead, these cache functions should just initialize those things it needs (that is the NSPR ssl error codes).

This patch does:
1) fixes the cache init code to only initialize error codes.
2) fixes the selfserv MP code to 1) be compatible with ssl.sh's selfserv management (at least on Unix), and 2) mimic the way real servers handle the MP_Cache init code (calling NSS_Init after the cache set up).
3) update ssl.sh server policy test to test policy usage on an MP server. This
is only done for non-windows like OS's because they can't catch the kill signal
to force their children to shutdown.

I've verified that the test fails if 2 and 3 are included but 1 is not
(and succeeds if all three are included).

Differential Revision: https://phabricator.services.mozilla.com/D70948