Bug 1629553 Prefix pk11wrap (SHLIB|LIBRARY)_VERSION with NSS_ r=rrelyea
In the manifest.mn the LIBRARY_VERSION is normally used to define
the major version of the build shared library. This ust works for
the pk11wrap case, because pk11wrap is a static library. But it's
still very confusing when reading the manifest.mn. Also the
referenced define in the code is just named SHLIB_VERSION.

So this prefixes the defines and the variables with NSS_, because
it tries to load the NSS library, just as the SOFTOKEN_.*_VERSION
is used to load the versioned softokn library.

Differential Revision: https://phabricator.services.mozilla.com/D70689